% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coords2spi.R
\name{coords2spi}
\alias{coords2spi}
\title{Convert coordinates to spatial information}
\usage{
coords2spi(points, global)
}
\arguments{
\item{points}{a data frame with two columns of coordinates, or a Spatial object listed here under "x" in showMethods(sp::over).
These represent points to be mapped onto polygon(s). Coordinates in `points` should be in the same column order as those in `global`}

\item{global}{an object onto which `points` is mapped, possible classes corresponding to `points` listed under "y" in
showMethods(sp::over). If `points` is a data frame, it will be converted into a SpatialPoints object.}
}
\value{
a data frame corresponding to the coordinates in `points` mapped onto `global` with the attributes of `global`
}
\description{
Wrapper function for `sp::over` in overlaying points over polygons to retrieve relevant intersection information.
}
\examples{
\dontrun{

require(rworldmap)
test <- demystas::coords2spi(as.data.frame(cbind(60,50)), getMap(resolution="low"))
}
}
\seealso{
\code{\link[sp]{over}}
}
\author{
Atreya Shankar
}
\concept{demystas}
