% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demystas-package.R
\docType{package}
\name{demystas-package}
\alias{demystas-package}
\alias{demystas}
\title{Demystifying Automation on Simple Tasks (demystas)}
\description{
There are a myriad of daily tasks which seem simple, but yet hide complex computational requirements.
These are the kind of challenges that we encounter that are not entirely part of our main objective,
but need to be overcome nonetheless. These include conducting complicated string searches, querying spatial
information and sometimes just organizing messy data.
}
\details{
This package aims to offer useful functions for data analysts
which might help in making these "on-the-fly" tasks lighter and more automated.
}
