% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grepsAbb.R
\name{grepsAbb}
\alias{grepsAbb}
\title{Sequential abbreviation mapping}
\usage{
grepsAbb(x, y)
}
\arguments{
\item{x}{a character vector containing abbreviations.}

\item{y}{a character vector containing whole words that could correspond to abbreviations.}
}
\value{
a matrix with total number of rows equal to the length of vector `x`. The first column is a repeat of vector `x`. Further columns represent matched possible `y` vector components.
}
\description{
Sequentially maps abbreviations onto complete words and returns a matrix of valid possibilities.
}
\examples{
\dontrun{

x <- c("BLG", "BLD", "LAT", "EMM")
y <- c("Boulder","Latino", "Eminem", "Emmys", "Building")
test <- demystas::grepsAbb(x, y)
}
}
\author{
Atreya Shankar
}
\concept{demystas}
