% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spNearest.R
\name{spNearest}
\alias{spNearest}
\title{Find nearest country (polygon) to point}
\usage{
spNearest(points, global, inc = 100)
}
\arguments{
\item{points}{a data frame with two columns of coordinates, or an object of class SpatialPoints or SpatialPointsDataFrame.}

\item{global}{an object of class SpatialPolygonsDataFrame onto which `points` is mapped.}

\item{inc}{a numerical value which indicates how much the entire bounding box of `global` shoud be segmented to find nearest countries. Defaults to 100.}
}
\value{
a data frame with attributes from `global` about nearest countries to `points`
}
\description{
Finds nearest country to a point, useful for cases where the `sp::over` function yield NA results
}
\examples{
\dontrun{

require(rworldmap)
require(rworldxtra)
points <- rbind(c(-81.779,52.234), c(-80.873, 51.126))
test <- spNearest(points, getMap(resolution="high"))
}
}
\author{
Atreya Shankar
}
