% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorEnum.R
\name{vectorEnum}
\alias{vectorEnum}
\alias{demystas}
\title{Enumerating vectors}
\usage{
vectorEnum(x, sep = ".")
}
\arguments{
\item{x}{a vector with possibly duplicated components}

\item{sep}{a character indicating how the enumeration and original vector components should be separated. Defaults to "."}
}
\value{
a vector in the same order and dimension as `x` with each component uniquely enumerated
}
\description{
Enumarates components of a vector in order to make each component unique. Useful for cases where duplicates of vector components must be avoided.
}
\examples{
\dontrun{

x <- c(rep("foo", 10), rep("bat", 25), rep("baz", 10), rep("foo", 10))
test <- vectorEnum(x)
}
}
\author{
Atreya Shankar
}
