% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coords2spi.R
\name{coords2spi}
\alias{coords2spi}
\title{Overlaying points onto polygons}
\usage{
coords2spi(points, global)
}
\arguments{
\item{points}{a data frame with two columns of coordinates, or an object of class SpatialPoints or SpatialPointsDataFrame.
These represent points to be mapped onto polygon(s).}

\item{global}{an object of class SpatialPolygonsDataFrame onto which `points` is mapped.}
}
\value{
a data frame corresponding to the coordinates in `points` mapped onto `global` with the attributes of `global`
}
\description{
Overlays coordinates of class data frame, SpatialPoints or SpatialPointsDataFrame onto polygon(s) of class SpatialPolygonsDataFrame
}
\examples{
\dontrun{

require(rworldmap)
test <- coords2spi(as.data.frame(cbind(60,50)), getMap(resolution="low"))
}
}
\author{
Atreya Shankar
}
