% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coords2spi.R
\name{coords2spi}
\alias{coords2spi}
\title{Convert coordinates to spatial information}
\usage{
coords2spi(points, global)
}
\arguments{
\item{points}{a data frame with two columns of coordinates, or an object of class SpatialPoints or SpatialPointsDataFrame.
These represent points to be mapped onto polygon(s). Coordinates in `points` should be in the same column order as those in `global`}

\item{global}{an object of class SpatialPolygonsDataFrame onto which `points` is mapped.}
}
\value{
a data frame corresponding to the coordinates in `points` mapped onto `global` with the attributes of `global`
}
\description{
Overlays coordinates of class data frame, SpatialPoints or SpatialPointsDataFrame onto polygon(s) of class SpatialPolygonsDataFrame in order to retrieve relevant intersecting
spatial information. Utilizes the `sp::over` function at the backend.
}
\examples{
\dontrun{

require(rworldmap)
test <- coords2spi(as.data.frame(cbind(60,50)), getMap(resolution="low"))
}
}
\seealso{
\code{\link[sp]{over}}
}
\author{
Atreya Shankar
}
\concept{demystas}
